+++ packages/sdks/core-js-sdk/test/sdk/helpers.test.ts
@@ -1,11 +1,11 @@
 import { isJwtExpired, pathJoin } from '../../src/sdk/helpers';
-import jwtDecode from 'jwt-decode';
-import {
-  MAX_POLLING_TIMEOUT_MS,
-  MIN_POLLING_INTERVAL_MS,
-} from '../../src/constants';
+import { jwtDecode } from 'jwt-decode';
 
-jest.mock('jwt-decode', () => jest.fn());
+jest.mock('jwt-decode', () => {
+  return {
+    jwtDecode: jest.fn(),
+  };
+});
 
 describe('helpers', () => {
   describe('pathJoin', () => {