+++ packages/sdks/react-sdk/test/components/Descope.test.tsx
@@ -262,4 +262,18 @@ describe('Descope', () => {
       );
     });
   });
+
+  it('should render web-component with store-last-authenticated-user', async () => {
+    render(
+      <AuthProvider projectId="project-1" storeLastAuthenticatedUser>
+        <Descope flowId="flow-1" />
+      </AuthProvider>,
+    );
+    await waitFor(() => {
+      expect(document.querySelector('descope-wc')).toHaveAttribute(
+        'store-last-authenticated-user',
+        "true",
+      );
+    });
+  });
 });