+++ packages/sdks/vue-sdk/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.0.32](https://github.com/descope/descope-js/compare/vue-sdk-2.0.31...vue-sdk-2.0.32) (2024-09-05)
+
+### Dependency Updates
+
+* `audit-management-widget` updated to version `0.1.84`
+## [2.0.31](https://github.com/descope/descope-js/compare/vue-sdk-2.0.30...vue-sdk-2.0.31) (2024-09-03)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.120`
+* `audit-management-widget` updated to version `0.1.83`
+* `role-management-widget` updated to version `0.1.118`
+* `user-management-widget` updated to version `0.4.121`
+* `user-profile-widget` updated to version `0.0.102`
+* `web-component` updated to version `3.24.1`
+* `web-js-sdk` updated to version `1.16.4`
+* `core-js-sdk` updated to version `2.24.4`
+## [2.0.30](https://github.com/descope/descope-js/compare/vue-sdk-2.0.29...vue-sdk-2.0.30) (2024-09-02)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.119`
+* `audit-management-widget` updated to version `0.1.82`
+* `role-management-widget` updated to version `0.1.117`
+* `user-management-widget` updated to version `0.4.120`
+* `user-profile-widget` updated to version `0.0.101`
+* `web-component` updated to version `3.24.0`
 ## [2.0.29](https://github.com/descope/descope-js/compare/vue-sdk-2.0.28...vue-sdk-2.0.29) (2024-08-20)
 
 ### Dependency Updates