+++ packages/widgets/role-management-widget/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.118](https://github.com/descope/descope-js/compare/role-management-widget-0.1.117...role-management-widget-0.1.118) (2024-09-03)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.16.4`
+## [0.1.117](https://github.com/descope/descope-js/compare/role-management-widget-0.1.116...role-management-widget-0.1.117) (2024-09-02)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.3.1`
 ## [0.1.116](https://github.com/descope/descope-js/compare/role-management-widget-0.1.115...role-management-widget-0.1.116) (2024-08-20)
 
 ### Dependency Updates